Fix various gcc warnings
variable-length array:

  ../src/smp.c:92:11: warning: ISO C forbids zero-size array ‘read_data’ [-Wpedantic]
  ../src/smp.c:121:11: warning: ISO C forbids zero-size array ‘data’ [-Wpedantic]

unused variables:

  ../src/block.c:101:46: warning: unused parameter ‘device’ [-Wunused-parameter]
  ../src/dellssd.c:234:36: warning: unused parameter ‘path’ [-Wunused-parameter]
  ../src/ledctl.c:131:36: warning: unused parameter ‘ignore’ [-Wunused-parameter]
  ../src/smp.c:544:34: warning: unused parameter ‘path’ [-Wunused-parameter]
  ../src/vmdssd.c:167:35: warning: unused parameter ‘path’ [-Wunused-parameter]

different signedness:

  ../src/config_file.c:87:11: warning: comparison of integer expressions of
  different signedness: ‘int’ and ‘long unsigned int’ [-Wsign-compare]
  ../src/vmdssd.c:156:37: warning: comparison of integer expressions of different
  signedness: ‘int’ and ‘size_t’ {aka ‘long unsigned int’} [-Wsign-compare]

binary constants:

  This x4:
  ../src/vmdssd.c:36:30: warning: binary constants are a GCC extension

variables to match sscanf data types:

  ../src/dellssd.c:254:27: warning: format ‘%x’ expects argument of type
  ‘unsigned int *’,
  but argument 3  has type ‘int *’ [-Wformat=]

comparison of fgets() return value:

  The fgets() returns a pointer, or NULL.  Following warning made me to notice
  this.

  ../src/smp.c:291:34: warning: ordered comparison of pointer with integer
  zero [-Wpedantic]

scanf type modifier:

  Lets use the modifier that matches with datatype, rather than casting and
  more problematic modifier.

  ../src/utils.c:107:18: warning: ISO C does not support the ‘%Lx’ gnu_scanf
  format [-Wformat=]

stop using not existing identifier:

  ../src/utils.c:523:40: warning: ISO C does not support ‘__FUNCTION__’
  predefined identifier [-Wpedantic]

variable shadowing warnings:

  ../src/ledctl.c:269:52: warning: declaration of ‘ibpi_list’ shadows a
  global declaration [-Wshadow]

  And same message about these as well.

  ../src/ledctl.c:282:57: ‘ibpi_list’
  ../src/ledctl.c:614:46: ‘ibpi_list’
  ../src/ledmon.c:172:69: ‘progname’
  ../src/ledmon.c:353:61: ‘longopt’
  ../src/scsi.c:421:15: ‘element_type’
  ../src/sysfs.c:565:44: ‘slave_list’

and old-style function definitions.

+++ src/utils.h
@@ -175,7 +175,7 @@ int scan_dir(const char *path, struct list *result);
  *
  * @return Number of bytes written if successful, otherwise -1 for an error.
  */
-int buf_write(const char *path, const char *buf);
+ssize_t buf_write(const char *path, const char *buf);
 
 /**
  * @brief Reads the content of a text file.
